Refactor observation weather analysis into a modular package
This commit decomposes the 482-line monolithic `apts/observations/weather.py`
module into a modular package structure:
- `constants.py`: Holds default values and data mappings.
- `engine.py`: Houses the complex vectorized analysis logic.
- `base.py`: Implements the `WeatherAnalysisMixIn` class.
- `__init__.py`: Maintains public API compatibility.

This improvement follows the Single Responsibility Principle and enhances
maintainability and readability of the observation weather suitability logic.
Internal test mocks were updated to reflect the new structure.
